Weather/weather_etl/requirements.txt
# Essential Data Libraries
pandas==2.1.4
numpy==1.26.2

# API Handling
requests==2.31.0
urllib3==2.1.0

# Database & SQL
psycopg2-binary==2.9.9    # PostgreSQL
mysql-connector-python==8.3.0    # MySQL
sqlalchemy==2.0.25

# Workflow Orchestration
apache-airflow==2.7.3

# Cloud & Storage (If needed)
boto3==1.34.23   # AWS S3
google-cloud-storage==2.14.0    # Google Cloud Storage

# Logging & Debugging
loguru==0.7.2
